requirements.txt
# Core document handling
python-docx>=1.1.0

# LLM
anthropic>=0.25.0

# Data models
pydantic>=2.0.0

# Web framework
flask>=3.0.0

# HTTP requests (search provider calls)
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Type-checking helpers (stdlib backport for older Python)
typing-extensions>=4.9.0
